Bug fix: activity outputs should be overriden when the activity is called multiple time.
In #69, we fixed a bug about activities outputs being overridden when an activity has multiple outputs to store. We implemented a sanity method called before writing outputs to check no outputs already exist  for that activity in the context. In this case, a runtime exception is thrown.
When we use loops, an activity is called multiple times and then, in the same process, the activity writes its outputs multiple times. This was not accepted anymore because of that sanity check. We removed it in this commit.

A small change is done in the WorkflowAssert: when an activity is executed multiple times, only the last execution outputs are kept in the context. But the method we use to test outputs use a history details which returns all outputs. We were getting the first occurence which does not correspond to the last execution. We replaced that with a reduce method to get the last one. THis allowed to test the last written outputs for an activity.