PLAT-11617: Expose util functions in execute-script activity
UtilFunctionMapper provides utility function that can be used in SWADL. So far, it was required to import them or to use to full path in order to call a utility method.
In this method, we made the UtilFunctionsMapper a bean and created a process engine plugin that injects it in Camunda context. We used utilityFunctionsMapper as a key in order to call method with utilityFunctionsMapper.escape(..) but this key can be changed to a better one if needed.
Doc and tests are updated following this change.