PLAT-11620: Support looping on single-replied form (#84)
The main problem is that form replies start subprocesses each one having its own life. But sometime in 1-1 or in specific case a form is meant to be replied by only one user. Meaning we can avoid using the subprocess and allow for loops to happen after the form is being replied to.

Introduce a flag in the form-reply event to decide wether subprocesses are used or not. The timeouts are not yet handled.

In this PR, we also added the support of timeouts in forms that require an exclusive reply. Forms reply timeout is implemented in the same way as in the other activities. If no timeout is provided, then the default PT24H is used.
In terms of implementation, the timeout for single reply forms is done in the same method as for the other activities. This allows us to maintain one method for both of them and we ensure having the same behaviour.


Co-authored-by: Youri Bonnaffe <youri.bonnaffe@symphony.com>