PLAT-11617: Expose util functions in execute-script activity (#85)
UtilFunctionsMapper provides utility function that can be used in SWADL. So far, it was required to import them or to use to full path in order to call a utility method.
In this method, we inject UtilFunctionsMapper in Camunda script engine. We used wdk as a key in order to call method with utilityFunctionsMapper.escape(..) but this key can be changed to a better one if needed.
Doc and tests are updated following this change.