Avoid overwriting activity outputs with form replies
Because we set the process variable under the same name (activity
id/form id) than the activity outputs for instance, we are overwriting
them when receiving a form reply.

Instead we now store them in a different name, just pass them with the
message that is triggering a form reply activity and with an execution
listener we properly merge them with existing variables like outputs.

This way we don't break the existing API.