Group support in the WDK
Groups (distribution lists) have recently been added in the BDK, we now
support them in the WDK too.

New activities:
- create-group
- update-group
- add-group-member
- get-group
- get-groups

Update avatar is supported through update-group to keep it simple.
The activities are pretty close to the APIs expect for some easy renames
to ease comprehension.

Also upgrade generator usage in doc.
Also upgrade dependencies.
Try to fix loop tests that are flaky because activities are executed too
fast so they have the same timestamp and are not properly ordered when
retrieved.