Added encodeRawQuery and decodeRawQuery to encode/decode url query parameters (#107)
Encode query parameters in execute-request.
In this PR, we introduced an optional field encode-query-params to execute-request activity in order to control whether we want the query parameters, given along with the full url, to be automatically encoded by the activity or not. It is by default set to True. 
We use Java standard URLEncoder to encode the query parameters in UTF-8 format.

We also made integration tests to wait at most 20 seconds checking executed activities in assertion, rather than 10 seconds that was not enough which made some tests fail.