#28: Support OBO (#112)
* #28: Support OBO
This PR is a first take on OBO support for WDK activities.
An optional property obo has been added in SWADL to make an activity call OBO, when this activity is OBO enabled.
In order to avoid making an authentication request for each OBO activity, we introduced a Caffeine cache to cache user obo authentication sessions. The cache is configured in application.yaml file.