Access control for monitoring apis (#144)
In this PR we added access control to the monitoring apis.
The monitoring apis require now a token "X-Monitoring-Token" in the request header in order to authorise the request.
The token value to match is provided as an environment variable to WDK jar and it is stored as a property in application.yaml with a placeholder. 

In terms of implementation, an annotation Authorized that contains the token key is added. This annotation is handled as a pointcut with AspectJ in order to match tokens.  
When no token is configured, all monitoring requests will be, by design, unauthorised.

For the same occasion, WDK config has been changed: the workflows folder property becomes under wdk.workflows.path and a new property wdk.properties.montoring-token is added for the monitoring api token.

For the same occasion, we enriched WorkflowBotConfiguration class with more logic to handle WDK configuration such as the workflows file and tokens.