Monitoring api: Support query parameters (#147)
In this PR, we added the support of query parameters to filter results.

The list workflow instances endpoint takes the status=pending/completed/failed as query parameter. An instance is considered as failing if the last executed activity is not the end event.
The list instance activities endpoint takes 4 possible query parameters started_before/started_after/finished_before/finished_after to filter activities in time.
The list global variables endpoint takes two query parameters occurred_before/occurred_after to filter variables changes in time.
More changes have been done in this PR:

When the monitoring token is not provided in the header, we return 401 instead of 400.
Fixed 500 error in the list global variables endpoints when the provided instance id is not correct.
Added nodeId in the activities definition, for both activities and events.
In the list of workflow instances and workflow activities endpoints, we added a sanity check to verify that the provided instanceId belongs to the provided workflow. Otherwise an error is thrown.
Added more tests