#149: Fix bug when on request-received event is added with a subsequent activity having form-replied event.
The issue happens when an activity with triggering event request-received within one-of property that sends a form, and a subsequent activity executed on that form replied. The workflow being built has this request-received event attached to an endEvent which makes the workflow completes once the request is received without executing the activities.
The issue comes from an incomplete condition checking if the activity is in a loop. We enriched this condition by checking if the current activity is of type ACTIVITY_COMPLETED_EVENT.