FIINOS-WDK 153: Fix bug where a message contains break/new lines characters raises an Illegal unquoted character exception.  (#157)
In CamundaExecutor, when we map the activity json string to th executor activity object using object mapper, if the string contains a break/new lines characters, then the Object mapper fails to read
the value and raises the error. The fix was to replace all occurences of new and break lines with an escaped new line character.