Fix NPE bug in a definition having a FormRepliedEvent
The bug is a regression following the change, in which we reverted the solution
for an exclusive FormRepliedEvent not using camunda subprocess.

The NPE happens on the timeout event coming along with a form reply event, this
extra event is automatically generated and was mising the event type.

The fix is to set the event type when this event is created.