#156 Enhance blast message and support multiple events  (#180)
Enhance blast message
So far, when sending a blast message, only the first messageId was stored in activity outputs. In this PR, we introduced two new outputs storing the list of message Ids, and the list of the failing ones.
For the same occasion, when a send-message activity has a list of user-ids, a blast message is sent to each user instead of creating a room (MIM) with all of them. User ids are converted to stream ids: a private message with the bot (IM) if exists, otherwise a new stream is created.

Support multiple events
So far, only the latest event being received was stored in variables and could be accessed with the key "event". This is due to the fact that, every time an event is received, a variable with the key "event" was created, therefore the previous event is overridden. In this PR, we changed this behaviour to store events with a unique key in order to keep all of them. For that, an optional id in introduced in SWADL for events, when it is set and the event is received, a new variable is stored. The event can be accessed in SWADL using this id. Please note that the "id" property is not unique and two events can use the same one, in this case the last one to be received will be stored. Fix this issue in a clean way would required a refactoring WDK code planned in a next PR.