Upgrade Camunda and h2 database (#186)
Upgraded H2 database version to the latest 2.1.214. This will fix CVE CVE-2022-23221.

Upgraded Camunda version to 7.18.0 in order to allow H2 upgrade as this H2 version was not compatible with Camunda 7.16.0. This latest version has introduced a breaking change in ExpressionManager class that becomes an interface. The method we were using to bind UtilityFunctions in ExpressionManager has changed. In this commit, I introduced UtilityFunctions class defining functions names and binded them in the expressionManager using Camunda ReflectUtils