Fix custom activity type unknown issue in definitino endpoint (#192)
The activity type and group were predefined in an enum class, that prevent
the definition endpoint from knowing a custom activity's type and group (with
null value).

This commit fix the issue. Instead of doing the mapping through the enum class,
now it is dynamically reading and converting from the activity class simple name.