#198 Expose Bot User Profile. (#205)
In this PR, we encriched the Utility functions with a new method allowing to retrieve the bot session info. For this purpose, a static sessionService field is added to the class, the SessionService bean is passed in the public constructor and set to the static field in order to be used in the static method session().