Workflow version control foundation (#206)
In this PR we introduced the version control for workflows.

Workflow versions are stored in a dedicated datasource managed with JPA. A workflow entity has a unique couple that consists of the workflow id and the version, the swadl is stored along with this entity as a blob. For future improvement, it would need to be compressed to save space.
An index is created on this unique couple as most of queries filter by workflowId or workflowId and version. 

WDK contains now 2 separate databases, one for Camunda to store the processes and a WDK dedicated one to store the swadls. Both are configurable in the application.properties file. They are memory based by default, so no persistence.

Some data that were stored in hashmaps for verifications is replaced by this datasource as well. A cache would be required to improve this part and avoid hitting the database for every verification.

An api endpoint is added to set a previous workflow version as the active one.

Some unrelated unit tests files have been updated for more test coverage.