Refactor to fix Audit trail logger (#208)
Audit trail logger was broken due to the introduction of WorkflowEventVariableAction. Both classes implement HistoricEventHandler from Camunda and the WorkflowEventVariableAction was set as primary bean. This leads to the issue where an historic event should be processed by both beans, only the primary one does. In this PR, a refactor is done to implement both beans in a Command pattern way, with an HistoricEventHandler who dispatches events to both beans that either process it or ignore it based on the event type.