Refactor the version control
Separate the two swadl management process, disk and management api.
Use version control when manage swadl through management api only.
This simplifies the complexity of the version control and the maintenance
in the future.

Version is controlled on server side and becomes mandatory in DB,
currently the version value is a timestamp at macro precision.

Reorganise the two datasource configuration in application.yml.

The DB used for version control is not required when the management api
is disabled (management token is not set).