Disable org.camunda.bpm log to avoid Camunda stacktrace (#219)
When an activity execution starts, the NodeBuilder stores its swadl as String in order to allow CamundaExecutor to read it and build the activity to be executed. The insert query is not executed on the fly but only after the execution ends, as a batch. When the activity swadl exceeds 4000 char, the insert query batch fails as the variable is of type String requiring a max of 4000 chars. This exception cannot be caught in WDK as it is executed by a Camunda scheduled job. The exception does not make the workflow to fail but only logs the stacktrace. As the variable only used by the CamundaExecutor, and it is not fetched after the activity execution, we decided to only hide it from the console by setting the log level of org.camunda.bpm to OFF.