Fix versions filter in the instances management api (#222)
Fix versions filter in the instances management api.
When a version parameter is provided, the list instances management api was not removing entries with unmatching versions, the converter was setting null as version for those unmatching instances. This is due to the fact that the processDefinitionKey, which is the workflow id, was used for filtering. This has been replaced by the processDefinitionId which is the id of the matching versions instances.