Fix response code when trying to update a published workflow. (#224)
The error message returned when trying to update a workflw that has an already published version was confusing at it says that the workflow does not exist. THis has been fixed in this PR by firstly checking if the workflos exists, then returning a more detailed exception when a version is already published.