[CICD] - Fixed Docker image push job in the release workflow. (#234)
The docker image push job depends on the release job but it needs to checkout the repository when it starts which make it lose the fatJar procuded by the previous job. In this commit, we added actions to the release job to download the fatJar as an artifact and download it in the next job to create the docker image.