Fix event listener method signature bug
Since BDK 2.13.0, the RealTimeEvent object is actually a sub type
of the original V4 event type, therefore change the Spring
event listener method signature to include the sub type of the event.

It is also a good practice to include the sub type to have maximum
extensibility in the code.