Fix issue to support multi timer fired event
A map is used to register the events and activities in the workflow
graph, timer fired event uses a fixed string as id, in case there is
several multi timers, the latest one overwrites the previous ones.

This commit fixed the issue by appending timer value to the event
name, so they should be unique most of time, except if the timer
value is the same (work around is to use all-of).