Fix activity variable size limit bug
The activity json string input variable might be too big
to be stored in DB (4000 characters limit), e.g. the activity
json string contains a variable being resolved at runtime with
a more than 4000 characters string value, when camunda persists
this new activity json string as variable in DB, it fails with
an exception.

This commit fixed this issue by definiting the input activity
string in a map instead of string, the map object is stored in
bytearray table as BLOB type without limit.